Restructured project before di implementation and package rename
