+++ app/src/main/java/com/example/myapplication/ui/exercise/ExerciseFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.fragments
+package com.example.myapplication.ui.exercise
 
 import android.os.Bundle
 import android.util.Log
@@ -6,22 +6,16 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageButton
-import android.widget.ProgressBar
 import android.widget.RadioButton
-import android.widget.TextView
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.example.myapplication.*
-import com.example.myapplication.adapters.ExerciseAdapter
-import com.example.myapplication.adapters.CalendarAdapter
+import com.example.myapplication.ui.exercise.calendar.CalendarAdapter
 import com.example.myapplication.database.ProgramsAndExercisesDatabase
-import com.google.android.flexbox.*
 import com.example.myapplication.R
 import com.example.myapplication.databinding.FragmentMainBinding
 
 
-class MainFragment : Fragment(), CalendarAdapter.ViewHolder.OnDateClick,
+class ExerciseFragment : Fragment(), CalendarAdapter.ViewHolder.OnDateClick,
     ExerciseAdapter.OnExerciseClick {
 
     private lateinit var binding: FragmentMainBinding
