+++ app/src/main/java/com/example/myapplication/ui/exercise/ExerciseViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.myapplication.ui.exercise
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+
+class ExerciseViewModel(application: Application) : AndroidViewModel(application) {
+}
