+++ app/src/main/res/navigation/app_nav.xml
@@ -7,12 +7,12 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.example.myapplication.fragments.MainFragment"
+        android:name="com.example.myapplication.ui.exercise.ExerciseFragment"
         android:label="fragment_main"
         tools:layout="@layout/fragment_main" />
     <fragment
         android:id="@+id/programFragment"
-        android:name="com.example.myapplication.fragments.ProgramFragment"
+        android:name="com.example.myapplication.ui.programs.ProgramFragment"
         android:label="fragment_program"
         tools:layout="@layout/fragment_program" />
     <fragment
